+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Upgraded client dependencies
+- Upgraded Arend to latest dev (API changed)
+
 ## 0.2.9
 
 - Upgraded client dependencies